% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetComparison.R
\docType{data}
\name{class.colors}
\alias{class.colors}
\title{DsFeatFreqComp – Dataset Feature-Frequency Comparison (Visualization)}
\format{An object of class \code{character} of length 8.}
\usage{
class.colors
}
\description{
Visualization of the comparisons among a group of datasets or
between the datasets in pairs. Especially, representing p-values.
}
\note{
version history
1.1, 16 Nov 2018, Batch processing
1.0, 21 Jun 2018, The first version
libraries
}
\references{
G. Canbek et al. (2018) Gaining New Insight in Datasets via Multiple Binary-Feature Frequency Ranks with A Benign/Malign Mobile Apps Example, Journal of Machine Learning Research (JMLR) (Submitted)
(\href{http://gurol.canbek.com/Publications/}{Publications by Gurol Canbek})
}
\author{
Gürol Canbek, <gurol44@gmail.com>
}
\keyword{analysis,}
\keyword{dataset}
\keyword{engineering}
\keyword{feature}
\keyword{quantitative}
