% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetComparison.R
\docType{data}
\name{dataset.colors}
\alias{dataset.colors}
\title{Example colors for datasets}
\format{An object of class \code{character} of length 9.}
\usage{
dataset.colors
}
\description{
Example colors for datasets
}
\keyword{datasets}
