% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dscompmeanranksnegative.R
\docType{data}
\name{dscompmeanranksnegative}
\alias{dscompmeanranksnegative}
\title{Pairwise Dataset Comparison of Mean Ranks (Benign Software)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 15 rows and 3 columns.}
\source{
\href{http://canbek.com/datasets}{Datasets by Gurol Canbek}
}
\usage{
data(dscompmeanranksnegative)
}
\description{
Data compiled from academic sources for "Comparing Datasets via Feature Ranks by Adapted Kruskal-Wallis Test: A Benign/Malign Mobile Apps Example" academic study by Gurol Canbek
}
\examples{
data(dscompmeanranksnegative)
}
\references{
G. Canbek et al. (2018) Gaining New Insight in Datasets via Multiple Binary-Feature Frequency Ranks with A Benign/Malign Mobile Apps Example, Journal of Machine Learning Research (JMLR) (Submitted)
(\href{http://gurol.canbek.com/Publications/}{Publications by Gurol Canbek})
}
\keyword{Android,}
\keyword{datasets,}
\keyword{malware}
\keyword{malware,}
\keyword{mobile}
