% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dscompmeanrankspositive.R
\docType{data}
\name{dscompmeanrankspositive}
\alias{dscompmeanrankspositive}
\title{Pairwise Dataset Comparison of Mean Ranks (Malware)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 21 rows and 3 columns.}
\source{
\href{http://canbek.com/datasets}{Datasets by Gurol Canbek}
}
\usage{
data(dscompmeanrankspositive)
}
\description{
Data compiled from academic sources for "Gaining New Insight in Datasets via Multiple Binary-Feature Frequency Ranks with A Benign/Malign Mobile Apps Example" academic study by Gurol Canbek
}
\examples{
data(dscompmeanrankspositive)
}
\references{
G. Canbek et al. (2018) Gaining New Insight in Datasets via Multiple Binary-Feature Frequency Ranks with A Benign/Malign Mobile Apps Example, Journal of Machine Learning Research (JMLR) (Submitted)
(\href{http://gurol.canbek.com/Publications/}{Publications by Gurol Canbek})
}
\keyword{Android,}
\keyword{datasets,}
\keyword{malware}
\keyword{malware,}
\keyword{mobile}
