% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsfeatfreqnegative.R
\docType{data}
\name{dsfeatfreqnegative}
\alias{dsfeatfreqnegative}
\title{Dataset Features and Feature Frequencies for Negative Class (Benign Software)}
\format{An object of class \code{data.frame} with 59 rows and 7 columns.}
\source{
\href{http://canbek.com/datasets}{Datasets by Gurol Canbek}
}
\usage{
data(dsfeatfreqnegative)
}
\description{
Data compiled from academic sources for "Comparing Datasets via Feature Ranks by Adapted Kruskal-Wallis Test: A Benign/Malign Mobile Apps Example" academic study by Gurol Canbek
}
\examples{
data(dsfeatfreqnegative)
}
\references{
G. Canbek et al. (2018) Gaining New Insight in Datasets via Multiple Binary-Feature Frequency Ranks with A Benign/Malign Mobile Apps Example, Journal of Machine Learning Research (JMLR) (Submitted)
(\href{http://gurol.canbek.com/Publications/}{Publications by Gurol Canbek})
}
\keyword{Android,}
\keyword{datasets,}
\keyword{malware}
\keyword{malware,}
\keyword{mobile}
