% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsManipulation.R
\name{getPairwiseDsPValueMatrix}
\alias{getPairwiseDsPValueMatrix}
\title{Get pairwise dataset P value matrix from multicomparison p-values}
\usage{
getPairwiseDsPValueMatrix(df_multicompare)
}
\arguments{
\item{df_multicompare}{data frame having 'DSi', 'DSj', 'PValue' columns}
}
\value{
matrix
}
\description{
Convert multi comparison data frame having 'DSi', 'DSj', 'PValue' columns
into a two dimensional matrix around PValue that can be used for
\code{\link{plotPairwiseDsPValuesHeatMap}} function
}
\examples{
data(dscompmeanrankspositive)
mat_pairwise_ds_pvalues <- getPairwiseDsPValueMatrix(dscompmeanrankspositive)
}
