% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsManipulation.R
\name{loadDsFeatFreqsFromCsv2}
\alias{loadDsFeatFreqsFromCsv2}
\title{DsFeatFreqComp – Dataset Feature-Frequency Comparison (Dataset Manipulation)}
\usage{
loadDsFeatFreqsFromCsv2(workbook, seperator = ";")
}
\arguments{
\item{workbook}{CSV file path}

\item{seperator}{Seperator character (default: ';')}
}
\value{
data frame having feature name in the first column and feature freq-
uencies for each datasets in subsequent columns (DSi)
}
\description{
Various dataset manipulation functions used in dataset comparison
load dataset feature frequencies from a CSV file

Read a CSV (comma seperated values) file having dataset feature frequencies
into a dataframe and return it.

CSV file should have the following columns and rows:
Feature_Name | DSA | DS0 | DS1 | ... | DSN
ABC | 0.1 | 0.3 | 0.999 | 0.76 | ... | 0.444
...
}
\references{
G. Canbek et al. (2018) Gaining New Insight in Datasets via Multiple Binary-Feature Frequency Ranks with A Benign/Malign Mobile Apps Example, Journal of Machine Learning Research (JMLR) (Submitted)
(\href{http://gurol.canbek.com/Publications/}{Publications by Gurol Canbek})
}
\author{
Gürol Canbek, <gurol44@gmail.com>
}
