% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsManipulation.R
\name{meltDataFrame}
\alias{meltDataFrame}
\title{Melt a dataframe having all the datasets' feature and frequencies into
feature, dataset and frequency}
\usage{
meltDataFrame(df, identity_columns = names(df)[1])
}
\arguments{
\item{df}{Data frame with Feature_name column and feature frequency per
dataset}

\item{identity_columns}{Column name for the feature name (default first
column's name)}
}
\value{
data frame
}
\description{
Melt (transform) a dataframe into the following structure (transactions) and
return it. \code{identity_columns} become \code{variable} and values under
each dataset column becomes 'value'.
}
\details{
Feature_Name | DSA | DS0 | DS1 | ... | DSN
ABC | 0.1 | 0.3 | 0.999 | 0.76 | ... | 0.444

into a feature per single dataset per value rows. For example

Feature_Name | variable | value
ABC | DSA | 0.1
... | DSA | ...
ABC | DS0 | 0.3
... | DS0 | ...
ABC | DS1 | 0.999
... | DS1 | ...
...
ABC | DSN | 0.444
... | DSN | ....
}
\note{
To change original column (dataset) order use \code{df[,c(1, 8, 2:7)]}
which moves the last column (i.e. dataset) into the second column.
}
\examples{
df_melted <- meltDataFrame(df)
}
