% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetComparison.R
\name{plotDsFreqDistributionViolin}
\alias{plotDsFreqDistributionViolin}
\title{plot dataset frequency distribution as a violin graph}
\usage{
plotDsFreqDistributionViolin(df_melted, condition = "Positive",
  ds_palette = dataset.colors, label_x = "Datasets",
  label_y = "Feature frequency probability densities, ranges, quartiles, means, outliers")
}
\arguments{
\item{df_melted}{Data frame melted as Feature_Name | variable (DSi) | value
(feature frequency for DSi)}

\item{condition}{Datasets' class (default: \code{'Positive'} or
\code{'Negative'})}

\item{ds_palette}{Dataset colors (default: \code{dataset.colors})}

\item{label_x}{X axis label (default: \code{'Datasets'})}

\item{label_y}{Y axis label (default: \code{'Feature frequency probability
densities, ranges, quartiles, means, outliers'})}
}
\value{
None
}
\description{
Plot feature frequency statistics per dataset as a violin graph (frequency
probability densities) and descriptive statistics: ranges (min/max values
shown in vertical line ends), quartiles (lower and upper shown in bottom and
top edges of box), medians (horizontal line in box), means (black dot), and
outliers (pink dot) per dataset.
}
\note{
Development notes:
}
\examples{
data(dsfeatfreqpositive)
df_melted <- meltDataFrame(dsfeatfreqpositive)
plotDsFreqDistributionViolin(df_melted)
}
\seealso{
http://www.sthda.com/english/rpkgs/ggpubr/
For pairwise comparison add
\code{my_comparisons <- list(c('DS0', 'DSA'), c('DS0', 'DS1'),
                             c('DS0', 'DS2'))}
Add
\code{ggpubr::stat_compare_means(comparisons=my_comparisons, label = 'p.signif')+}
}
