% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetComparison.R
\name{plotPairwiseDsPValuesHeatMap}
\alias{plotPairwiseDsPValuesHeatMap}
\title{Plot pairwise dataset p-values as a heatmap diagram}
\usage{
plotPairwiseDsPValuesHeatMap(mat_pairwise_ds_pvalues)
}
\arguments{
\item{mat_pairwise_ds_pvalues}{Pairwise dataset P values}
}
\value{
None
}
\description{
Plot a heatmap from two dimensional p-value matrix for datasets. The colormap
is originally arranged to indicate the rejection/acceptance limits for null
hypothesis stating the similarity:
}
\details{
Reject: 0, 0.001, 0.01, 0.05; Failed to Reject: >> 0.05
}
\note{
Some resources:
\url{https://stackoverflow.com/questions/24919413/selecting-number-of-leaf-nodes-of-dendrogram-in-heatmap-2-in-r}
\url{https://talesofr.wordpress.com/2013/05/05/ridiculously-photogenic-factors-heatmap-with-p-values/}
}
\examples{
data(dscompmeanrankspositive)
mat_pairwise_ds_pvalues <- getPairwiseDsPValueMatrix(dscompmeanrankspositive)
plotPairwiseDsPValuesHeatMap(mat_pairwise_ds_pvalues)
}
