% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetComparison.R
\name{plotQQ}
\alias{plotQQ}
\title{Plot Quantile-Quantile chart for all the dataset in one figure}
\usage{
plotQQ(df, condition = "Positive", ds_palette = dataset.colors,
  skip_column = 1, ds_names = NULL, label_x = FALSE)
}
\arguments{
\item{df}{Data frame with Feature_name column and feature frequency per
dataset}

\item{condition}{Datasets' class (default: \code{'Positive'} or
\code{'Negative'})}

\item{ds_palette}{Dataset colors (default: \code{dataset.colors})}

\item{skip_column}{Number of columns skipped (default: 1 [feature name
column])}

\item{ds_names}{Name of the datasets (default: \code{NULL} [retrieve from
dataframe])}

\item{label_x}{X axis label (default: \code{FALSE} [Add Shapiro normality test])}
}
\value{
none
}
\description{
Plot Quantile-Quantile chart with Shapiro-Wilk test values and P-values are
provided for each dataset. If some of the frequency values (points) are
outside the corresponding normal distribution indicated by a shaded area it
means that the dataset does not have a normal distribution. All the datasets
are plotted into side-by-side in two columns
}
\examples{
library(dsfeatfreqcomp)

# Colors for datasets
dataset.colors <- ggsci::pal_simpsons()(9)
# Swap colors to comply DSA, DS0, DS1, ... order
dataset.colors[c(2,3)] <- dataset.colors[c(3,2)]
dataset.colors[c(1,2)] <- dataset.colors[c(2,1)]

data(dsfeatfreqnegative)
# There is no negative sample for DS5 arrange colors
plotQQ(dsfeatfreqnegative, 'Negative', ds_palette=dataset.colors[c(1:5, 7)])

data(dsfeatfreqpositive)
# For single dataset plot only (y label is normally 'Sample'
# x label is 'Theoretical' in single ggqqplot)
plotQQ(as.data.frame(dsfeatfreqpositive[,2:2]), skip_column= 0, ds_names = 'DS2',
label_x='Theoretical')
}
\seealso{
\url{http://www.sthda.com/english/rpkgs/ggpubr/reference/ggqqplot.html}
}
