% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsProxy.R
\name{pressEnterToContinue}
\alias{pressEnterToContinue}
\title{Proxy Utulity Scripts
Stop script run and show a (custom) message to user to press ENTER

Show a given message with 'Press [enter] to continue' statement and wait for
the user interaction. It is useful for pausing script run}
\usage{
pressEnterToContinue(message = "")
}
\arguments{
\item{message}{custom message text to display}
}
\description{
Proxy Utulity Scripts
Stop script run and show a (custom) message to user to press ENTER

Show a given message with 'Press [enter] to continue' statement and wait for
the user interaction. It is useful for pausing script run
}
\references{
G. Canbek et al. (2018) Gaining New Insight in Datasets via Multiple Binary-Feature Frequency Ranks with A Benign/Malign Mobile Apps Example, Journal of Machine Learning Research (JMLR) (Submitted)
(\href{http://gurol.canbek.com/Publications/}{Publications by Gurol Canbek})
}
\author{
Gürol Canbek, <gurol44@gmail.com>
}
