% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetComparison.R
\docType{data}
\name{pvalue.colors}
\alias{pvalue.colors}
\title{Colors for p-values
Narrowed left colors original implementation by Gurol Canbek
Copyright (C) 2017-2018 Gürol Canbek
c('#C6EFCE', '#006100', # Green
  '#FFC7CE', '#9C0006', # Red
  '#FFEB9C', '#9C5700', # Orange
  'gray', 'black'       # Black
)}
\format{An object of class \code{character} of length 193.}
\usage{
pvalue.colors
}
\description{
Colors for p-values
Narrowed left colors original implementation by Gurol Canbek
Copyright (C) 2017-2018 Gürol Canbek
c('#C6EFCE', '#006100', # Green
  '#FFC7CE', '#9C0006', # Red
  '#FFEB9C', '#9C5700', # Orange
  'gray', 'black'       # Black
)
}
\keyword{datasets}
