% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetComparison.R
\docType{data}
\name{pvalues.breaks}
\alias{pvalues.breaks}
\title{Breaks for p-values
Narrowed left colors original implementation by Gurol Canbek
Copyright (C) 2017-2018 Gürol Canbek}
\format{An object of class \code{numeric} of length 194.}
\usage{
pvalues.breaks
}
\description{
Breaks for p-values
Narrowed left colors original implementation by Gurol Canbek
Copyright (C) 2017-2018 Gürol Canbek
}
\keyword{datasets}
