% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetComparison.R
\name{testCompareDsFeatFreq}
\alias{testCompareDsFeatFreq}
\title{Test comparison of datasets}
\usage{
testCompareDsFeatFreq(df, condition, palette)
}
\arguments{
\item{df}{Data frame with Feature_name column and feature frequency per
dataset}

\item{condition}{Datasets' class (default: \code{'Positive'} or
\code{'Negative'})}

\item{palette}{Dataset colors (default: \code{dataset.colors})}
}
\value{
None
}
\description{
Test (demonstration) for plotting Quantile-Quantile chart and Kruskal-Wallis
Violin chart for positive/negative class
}
\examples{
data(dsfeatfreqnegative)
# testCompareDsFeatFreq(dsfeatfreqnegative, 'Negative', dataset.colors[c(1:5, 7)])
testCompareDsFeatFreq(dsfeatfreqnegative, 'Negative', NULL)
}
