% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetComparison.R
\name{testCompareDsPValueHeatmap}
\alias{testCompareDsPValueHeatmap}
\title{Test heatmap comparison visualization for positive datasets}
\usage{
testCompareDsPValueHeatmap(df_multicompare)
}
\arguments{
\item{df_multicompare}{two dimensional matrix around PValue}
}
\value{
None
}
\description{
Test (demonstration) for plotting a heatmap from two dimensional P value
matrix for datasets with positive/negative class.
}
\examples{
data(dscompmeanrankspositive)
testCompareDsPValueHeatmap(dscompmeanrankspositive)
}
