% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetComparison.R
\name{testCompareNegativeDsFeatFreq}
\alias{testCompareNegativeDsFeatFreq}
\title{Test comparison of negative class datasets}
\usage{

  testCompareNegativeDsFeatFreq(csv = "Datasets/Negative_SeperatorSemiColon.csv")
}
\arguments{
\item{csv}{CSV file path}
}
\value{
None
}
\description{
Test (demonstration) for plotting Quantile-Quantile chart and Kruskal-Wallis
Violin chart for negative class
}
\examples{
# testCompareNegativeDsFeatFreq(csv='negative.csv')
}
