% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetComparison.R
\name{testCompareNegativeDsPValueHeatmap}
\alias{testCompareNegativeDsPValueHeatmap}
\title{Test heatmap comparison visualization for negative datasets}
\usage{

  testCompareNegativeDsPValueHeatmap(wb = "Datasets/NegativeDSMeanDifferencePValues.xlsx")
}
\arguments{
\item{wb}{Excel workbook path}
}
\value{
None
}
\description{
Test (demonstration) for plotting a heatmap from two dimensional P value
matrix for datasets with negative class.
}
\examples{
# testCompareNegativeDsPValueHeatmap(wb='NegativeDSMeanDifferencePValues.xlsx')
}
