% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetComparison.R
\name{testComparePositiveDsFeatFreq}
\alias{testComparePositiveDsFeatFreq}
\title{Test comparison of positive class datasets}
\usage{

  testComparePositiveDsFeatFreq(csv = "Datasets/Positive_SeperatorSemiColon.csv")
}
\arguments{
\item{csv}{CSV file path}
}
\value{
None
}
\description{
Test (demonstration) for plotting Quantile-Quantile chart and Kruskal-Wallis
Violin chart for positive class
}
\examples{
# testComparePositiveDsFeatFreq(csv='positive.csv')
}
