% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetComparison.R
\name{testComparePositiveDsPValueHeatmap}
\alias{testComparePositiveDsPValueHeatmap}
\title{Test heatmap comparison visualization for positive datasets}
\usage{

  testComparePositiveDsPValueHeatmap(wb = "Datasets/PositiveDSMeanDifferencePValues.xlsx")
}
\arguments{
\item{wb}{Excel workbook path}
}
\value{
None
}
\description{
Test (demonstration) for plotting a heatmap from two dimensional P value
matrix for datasets with positive class.
}
\examples{
# testComparePositiveDsPValueHeatmap(wb='PositiveDSMeanDifferencePValues.xlsx')
}
